requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
